+++ h2gis-functions/src/test/java/org/h2gis/functions/spatial/SpatialFunction2Test.java
@@ -914,7 +914,7 @@ public class SpatialFunction2Test {
     @Test
     public void test_ST_SVF3() throws Exception {
         Statement st = connection.createStatement();
-        ResultSet rs = st.executeQuery("SELECT ST_svf('POINT(0 0 0)'::GEOMETRY, ST_UPDATEZ(ST_FORCE3D(ST_buffer('POINT(0 0)'::GEOMETRY, 10, 120)), 12), 50, 8) as result");
+        ResultSet rs = st.executeQuery("SELECT ST_svf('POINT(0 0 0)'::GEOMETRY, ST_FORCE3D(ST_buffer('POINT(0 0)'::GEOMETRY, 10, 120), 12), 50, 8) as result");
         assertTrue(rs.next());
         double svfTest = 0.4098;
         assertEquals(svfTest, rs.getDouble(1), 0.01);
@@ -970,7 +970,7 @@ public class SpatialFunction2Test {
         Statement st = connection.createStatement();
         ResultSet rs = st.executeQuery(
                 "SELECT ST_svf('SRID=2154;POINT(0 0 0)'::GEOMETRY, " +
-                        "ST_UPDATEZ(ST_FORCE3D(ST_buffer('SRID=2154;POINT(0 0)'::GEOMETRY, 10, 120)), 12), 50, 8) as result");
+                        "ST_FORCE3D(ST_buffer('SRID=2154;POINT(0 0)'::GEOMETRY, 10, 120), 12), 50, 8) as result");
         assertTrue(rs.next());
         double svfTest = 0.4098;
         assertEquals(svfTest, rs.getDouble(1), 0.01);
@@ -980,7 +980,7 @@ public class SpatialFunction2Test {
     public void test_ST_SVF10() throws Exception {
         assertThrows(SQLException.class, () -> {
             st.execute("SELECT ST_svf('SRID=27572;POINT(0 0 0)'::GEOMETRY, " +
-                    "ST_UPDATEZ(ST_FORCE3D(ST_buffer('SRID=2154;POINT(0 0)'::GEOMETRY, 10, 120)), 12), 50, 8) as result");
+                    "ST_FORCE3D(ST_buffer('SRID=2154;POINT(0 0)'::GEOMETRY, 10, 120), 12), 50, 8) as result");
         });
     }