+++ docs/CHANGELOG.md
@@ -12,3 +12,4 @@ eg GEOMETRY(POINT) instead of only GEOMETRY
 + Update H2 from 2.0.204 to 2.0.206
 + Add a new module to run sql script tests (disable by default)
 + Update ST_Force3D to be inline with PostGIS
++ Update ST_UpdateZ to force the dimension when the z value is updated.